session_4/telegram_bot/pyproject.toml
[project]
name = "sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.1",
    "langchain-community>=0.3.30",
    "langchain-openai>=0.3.33",
    "langchain-pymupdf4llm>=0.4.1",
    "python-telegram-bot>=22.5",
    "reportlab>=4.4.4",
    "rich>=14.1.0",
]
